refactor: extract vcard parsing helper
